Factored out the HTTP relay client as a separate class to imrpvoe the ability to test the code.
