+++ src/go/cmd/http-relay-client/client/BUILD
@@ -1,18 +1,17 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library", "go_test")
-load("@rules_oci//oci:defs.bzl", "oci_image")
-load("@rules_pkg//pkg:tar.bzl", "pkg_tar")
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 package(default_visibility = ["//visibility:public"])
 
 go_library(
     name = "go_default_library",
-    srcs = ["main.go"],
-    importpath = "github.com/googlecloudrobotics/core/src/go/cmd/http-relay-client",
-    visibility = ["//visibility:private"],
+    srcs = [
+        "client.go",
+    ],
+    importpath = "github.com/googlecloudrobotics/core/src/go/cmd/http-relay-client/client",
     deps = [
-        "//src/go/cmd/http-relay-client/client:go_default_library",
         "//src/proto/http-relay:go_default_library",
         "@com_github_cenkalti_backoff//:go_default_library",
+        "@com_github_golang_glog//:go_default_library",
         "@io_opencensus_go//plugin/ochttp:go_default_library",
         "@io_opencensus_go//plugin/ochttp/propagation/tracecontext:go_default_library",
         "@io_opencensus_go//trace:go_default_library",
@@ -24,19 +23,18 @@ go_library(
     ],
 )
 
-go_binary(
-    name = "http-relay-client-app",
+go_test(
+    name = "go_default_test",
+    srcs = [
+        "client_test.go",
+    ],
     embed = [":go_default_library"],
-)
-
-pkg_tar(
-    name = "http-relay-client-image-layer",
-    srcs = [":http-relay-client-app"],
-)
-
-oci_image(
-    name = "http-relay-client-image",
-    base = "@distroless_base",
-    entrypoint = ["/http-relay-client-app"],
-    tars = [":http-relay-client-image-layer"],
+    visibility = ["//visibility:private"],
+    deps = [
+        "//src/proto/http-relay:go_default_library",
+        "@com_github_getlantern_httptest//:go_default_library",
+        "@com_github_onsi_gomega//:go_default_library",
+        "@in_gopkg_h2non_gock_v1//:go_default_library",
+        "@org_golang_google_protobuf//proto:go_default_library",
+    ],
 )
